requirements.txt
beautifulsoup4
requests
numpy
pandas
faiss-cpu
sentence-transformers
transformers
accelerate
spacy
tqdm
python-telegram-bot